identity-scrubber/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
Deprecated==1.3.1
fonttools==4.62.1
freetype-py==2.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
ollama==0.6.1
packaging==26.0
pdfminer.six==20260107
pikepdf==10.5.1
pillow==12.2.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.2.2
pypdfium2==5.7.0
pytesseract==0.3.13
socksio==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
wrapt==2.1.2
